requirements.txt
azure-core==1.38.2
azure-storage-blob==12.28.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
gunicorn==25.1.0
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pycparser==3.0
python-dotenv==1.2.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
